Refactor reader configuration loading to remove redundant code
