+++ satpy/scene.py
@@ -25,7 +25,8 @@ from satpy.config import get_environ_config_dir
 from satpy.dataset import (DataQuery, DataID, dataset_walker,
                            replace_anc, combine_metadata)
 from satpy.node import DependencyTree
-from satpy.readers import DatasetDict, load_readers
+from satpy.readers import load_readers
+from satpy import DatasetDict
 from satpy.resample import (resample_dataset,
                             prepare_resampler, get_area_def)
 from satpy.writers import load_writer
