+++ satpy/tests/test_scene.py
@@ -1857,7 +1857,7 @@ class TestSceneLoading(unittest.TestCase):
         """Test Scene available composites when composites don't have a dependency."""
         from satpy.tests.utils import FakeReader, test_composites
         import satpy.scene
-        from satpy.readers import DatasetDict
+        from satpy import DatasetDict
 
         def _test(self, sensor_names):
             if not self.compositors:
